test(config): split server/config.test.ts into three files and drop the process.env helper (#868)
server/config.test.ts was 910 lines against a max-lines rule of 500, and its
withEnv helper mutated process.env in violation of node/no-process-env. Both
findings meant the pre-commit gate refused every edit to the file, so any
future change to configuration tests was blocked on this split.

The file is split along its existing describe boundaries into three:

- server/config.test.ts keeps the boundary, optional-secret, and
  EMBEDDING_API_KEY blocks.
- server/config-extended.test.ts takes the extendedConfig helper and the
  extended env group blocks.
- server/config-equivalence.test.ts takes the start-equivalence reader
  comparisons and the shared-namespace helper block.

The fourth block no longer touches process.env. parseServerConfig is the pure
reader that loadServerConfig feeds process.env into, so a local parseFrom
helper builds the same environment shape as a plain object, calls
parseServerConfig, and throws a server_configuration_invalid error naming the
failing field when the parse fails. The four tests keep their assertions
unchanged, including the toThrow match on DB_HOST.

Every test and assertion moves verbatim. No test file reads or writes
process.env after this change.

Closes #868
